% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mda_ivm_functions.R
\name{ivm_fun_stag}
\alias{ivm_fun_stag}
\title{Function that returns parameters required for Ivermectin simulation with staggered distribution and metapop...}
\usage{
ivm_fun_stag(
  IVM_start_times_1,
  IVM_start_times_2,
  IVM_start_times_3,
  time_period,
  hazard_profile,
  ivm_coverage = 0.8,
  ivm_min_age = 5,
  ivm_max_age = 200,
  bites_per_3_days = 1
)
}
\description{
Function that returns parameters required for Ivermectin simulation with staggered distribution and metapop...
}
