% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mda_ivm_functions.R
\name{ivm_fun_stag_cov}
\alias{ivm_fun_stag_cov}
\title{Function that returns parameters required for Ivermectin simulation with staggered distribution and coverage}
\usage{
ivm_fun_stag_cov(
  IVM_start_times,
  time_period,
  hazard_profile,
  prop_human_HR_threshold,
  ivm_coverage = 0.8,
  ivm_min_age = 5,
  ivm_max_age = 200,
  bites_per_3_days = 1
)
}
\description{
Function that returns parameters required for Ivermectin simulation with staggered distribution and coverage
}
