% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mda_ivm_functions.R
\name{mda_fun}
\alias{mda_fun}
\title{Function that returns parameters required for MDA}
\usage{
mda_fun(
  ncc = 2,
  MDA_times = NULL,
  MDA_grp_prop = 0.9,
  MDA_cov = 0.8,
  MDA_st_age = 0.5,
  MDA_en_age = 80,
  MDA_succ = 0.95,
  MDA_drug_choice = 2
)
}
\description{
Function that returns parameters required for MDA
}
