% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{run_model}
\alias{run_model}
\title{run_model}
\usage{
run_model(
  age = c(0, 0.25, 0.5, 0.75, 1, 1.25, 1.5, 1.75, 2, 3.5, 5, 7.5, 10, 15, 20, 30, 40, 50,
    60),
  EIR = 10,
  ft = 0.4,
  admin2 = "Tororo",
  time = 365
)
}
\arguments{
\item{age}{Vector of age brackets.
Default=c(0,0.25,0.5,0.75,1,1.25,1.5,1.75,2,3.5,5,7.5,10,15,20,30,40,50,60)}

\item{EIR}{Numeric of annual EIR. Default=10}

\item{ft}{Numeric of proportion of symptomatic cases recieving treatment.
Default=0.4}

\item{admin2}{Character of admin unit. Default = "Tororo"}

\item{time}{Numeric of length of time that the model will run for in days.
Default=365}
}
\description{
\code{run_model} runs model using declared age, EIR, ft, country, admin
}
